requirements.txt
awsebcli==3.18.2
botocore==1.15.49
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
future==0.16.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
pathspec==0.5.9
PyMySQL==0.10.0
python-dateutil==2.8.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==1.0.1
python-dotenv==0.14.0
requests==2.20.1
authlib==0.15.2
six==1.11.0
